% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_total_saturation.R
\name{clip_vector}
\alias{clip_vector}
\title{Clip a vector from [a,b] to [0,1]}
\usage{
clip_vector(m, thr_low = 0, thr_high = 255, val_low = 0, val_high = 255)
}
\description{
Clip a vector from [a,b] to [0,1]
}
