% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_color.R
\name{detect_color}
\alias{detect_color}
\title{Detect if one image is in color or B&W}
\usage{
detect_color(img, ...)
}
\arguments{
\item{img}{a `magic-image` as returned by `[render_page()]`}

\item{...}{
  Arguments passed on to \code{\link[=get_total_saturation]{get_total_saturation}}
  \describe{
    \item{\code{sat_threshold_low}}{pixels with saturation lower than this value (from 0 to 255) will be clipped to 0}
    \item{\code{sat_threshold_high}}{pixels with saturation higher than this value (from 0 to 255) will be clipped to 255}
    \item{\code{thr_erode}}{morphologically erode the saturation before thresholding with a square kernel of side `thr_erode`}
  }}
}
\value{
a logical vector, as many frames as `img`
}
\description{
Detect if one image is in color or B&W
}
\examples{

img <- magick::wizard()

detect_color(img)

# The saturated areas
magick::wizard() \%>\% extract_saturation()

}
