% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate_pages.R
\name{enumerate_pages}
\alias{enumerate_pages}
\title{Enumerate pages and sheets in a pdf}
\usage{
enumerate_pages(pdf, double_sided)
}
\arguments{
\item{pdf}{path}

\item{double_sided}{if TRUE, sheets are printed on the front and back}
}
\value{
a tibble with columns:
- page: page number
- sheet: sheet number
- side: "front" or "back"
}
\description{
Enumerate pages and sheets in a pdf
}
