% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_pdf.R
\name{example_pdf}
\alias{example_pdf}
\title{An example PDF with some color and B&W pages}
\usage{
example_pdf()
}
\value{
path to a PDF file
}
\description{
Page index:
}
\details{
\enumerate{
\item color (vector)
\item B&W (vector)
\item (blank)
\item B&W (vector)
\item color (image)
\item B&W (image, color antialias)
\item color (vector)
\item color (vector)
\item color (vector)
\item B&W  (image, true B&W)
}
}
\examples{
example_pdf()
}
