% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_total_saturation.R
\name{get_total_saturation}
\alias{get_total_saturation}
\title{Summarize saturation values from an image}
\usage{
get_total_saturation(
  img,
  sat_threshold_low = 5,
  sat_threshold_high = 255,
  thr_erode = 0
)
}
\arguments{
\item{img}{a `magic-image` as returned by `[render_page()]`}

\item{sat_threshold_low}{pixels with saturation lower than this value (from 0 to 255) will be clipped to 0}

\item{sat_threshold_high}{pixels with saturation higher than this value (from 0 to 255) will be clipped to 255}

\item{thr_erode}{morphologically erode the saturation before thresholding with a square kernel of side `thr_erode`}
}
\value{
a numeric vector with the same length as `img`
}
\description{
Steps:
}
\details{
- extract the saturation channel using [extract_saturation()]
- apply a morphological erosion if asked
- round saturation values when too low/high
- sum all values in each image
}
\examples{
example_pdf() \%>\%
   render_page() \%>\%
   get_total_saturation()

}
