% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_pdf_chunks.R
\name{split_pdf_chunks}
\alias{split_pdf_chunks}
\title{Split a PDF to a color PDF and a B&W PDF}
\usage{
split_pdf_chunks(
  df_pages,
  pdf,
  filename_output = "{output_dir}/{filename}_{color}.pdf",
  output_dir = NULL
)
}
\arguments{
\item{df_pages}{tidy sheet dataframe, as returned by `[tidy_pdf_pages()]`}

\item{pdf}{input PDF to split}

\item{filename_output}{a template for the output PDF files}

\item{output_dir}{if not NULL, the directory that will contain the PDFs}
}
\value{
paths to color-separated PDFs
}
\description{
Split a PDF to a color PDF and a B&W PDF
}
\examples{

f <- example_pdf()

# Get sheet/color information
df_pages <- f \%>\%
   tidy_pdf_pages(double_sided = TRUE)

# Output files here
dir_temp <- tempdir()

# Split
f_split <- df_pages \%>\%
   split_pdf_chunks(
      pdf = f,
      output_dir = dir_temp
   )

print(f_split)
\dontrun{
fs::file_show(dir_temp)
}

}
