% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_pdf_pages.R
\name{tidy_pdf_pages}
\alias{tidy_pdf_pages}
\title{Extract page, sheet, color information from a PDF}
\usage{
tidy_pdf_pages(
  pdf,
  double_sided,
  density = NULL,
  opts_detect = list(),
  force_color = NULL,
  force_BW = NULL
)
}
\arguments{
\item{pdf}{path}

\item{double_sided}{if TRUE, sheets are printed on the front and back}

\item{density}{resolution to render pdf or svg}

\item{opts_detect}{list of parameters for [detect_color()]}

\item{force_color}{if not NULL, a vector of page numbers to be printed in color}

\item{force_BW}{if not NULL, a vector of page numbers to be printed in B&W}
}
\value{
a tibble with columns:
- page: page number
- sheet: sheet number
- side: the side of the page in the sheet: "front" or "back"
- img: a {magick} image
- is_color_page: TRUE if the page is in color
- is_color_sheet: TRUE if a page of the sheet is in color
}
\description{
Extract page, sheet, color information from a PDF
}
\examples{

f <- example_pdf()
f \%>\%
   tidy_pdf_pages(double_sided = TRUE)

f \%>\%
   tidy_pdf_pages(double_sided = TRUE, force_color = c(1, 2, 5))

}
